+++ internal/web/templates/participant_dashboard_templ.go
@@ -260,7 +260,7 @@ func Dashboard(p *model.Participant, module *model.Module, publicFiles []*model.
 					return templ_7745c5c3_Err
 				}
 			}
-			templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 23, "</div><span class=\"text-body-medium text-on-surface mb-2 block\">Modul aktif: <span id=\"module_name\">")
+			templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 23, "</div><span class=\"text-body-medium text-on-surface mb-2 hidden\"><span id=\"module_name\">")
 			if templ_7745c5c3_Err != nil {
 				return templ_7745c5c3_Err
 			}
@@ -338,7 +338,7 @@ func Dashboard(p *model.Participant, module *model.Module, publicFiles []*model.
 			}
 			return nil
 		})
-		templ_7745c5c3_Err = GuestLayout("Dashboard", "", p.Name).Render(templ.WithChildren(ctx, templ_7745c5c3_Var2), templ_7745c5c3_Buffer)
+		templ_7745c5c3_Err = GuestLayout("Dashboard").Render(templ.WithChildren(ctx, templ_7745c5c3_Var2), templ_7745c5c3_Buffer)
 		if templ_7745c5c3_Err != nil {
 			return templ_7745c5c3_Err
 		}