+++ internal/web/templates/scoring_templ.go
@@ -306,7 +306,7 @@ func Scoring(comp *model.Competition, participants []*model.Participant, modules
 							if templ_7745c5c3_Err != nil {
 								return templ_7745c5c3_Err
 							}
-							templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 21, "\" class=\"score-input input w-20 text-center\"></td>")
+							templ_7745c5c3_Err = templruntime.WriteString(templ_7745c5c3_Buffer, 21, "\" class=\"score-input input w-16 text-center px-2\"></td>")
 							if templ_7745c5c3_Err != nil {
 								return templ_7745c5c3_Err
 							}